Clean up some compiler warnings and prepare for wxWidgets 3.1

+++ src/wxsvg/src/SVGCanvasItem.cpp
@@ -543,7 +543,7 @@ double AngleBisect(float a1, float a2) {
 		delta += 2 * M_PI;
 	}
 	/* delta is now the angle from a1 around to a2, in the range [0, 2*M_PI) */
-	float r = a1 + delta / 2;
+	double r = a1 + delta / 2;
 	if (delta >= M_PI) {
 		/* the arc from a2 to a1 is smaller, so use the ray on that side */
 		r += M_PI;
